Make the tree controls in the Raw View and Plot View consistent (#57)
* Add Ctrl-A/Ctrl-C handling in plot view as in raw view

* Make the tree controls in the Raw View and Plot View behave the same

Ctrl-A expands and selects all items. Ctrl-C copies all items.

Previously, there was an issue with Ctrl-A in the raw view where
if you highlighted an unexpanded node with children and hit Ctrl-A,
the currently highlighted node would not itself expand. Now, all
nodes expand and are highlighted.

* Add treeview headers in raw view and plot view

It doesn't take up much space and is a convenient reminder
about the message type.

* Address code review feedback

Co-authored-by: Michael Jeronimo <michael.jeronimo@intel.com>