Fix a bug with the status line progress bar
When loading a bag with lots of topics, the progress bar was not
getting reset to 0 after the load completed. This was due to the
'updated' flag getting overwritten in the index caching thread and
not properly detecting that something had changed with the timeline.
If the last topic evaluated had no entries then it was erroneously
assumed that nothing at all had changed (for any topic).

Signed-off-by: Michael Jeronimo <michael.jeronimo@openrobotics.org>