Fix issue: no vertical scroll bar until window is resized
Previously, when loading a bag, if the new timeline height was larger
than the bag widget window height, the scroll bar would not show up until
the window was resized by the user. This fix forces a resize upon
opening a bag so that the new timeline height is taken into account
(now greater than the bag widget height) and the scrollbar appears,
if necessary.

Signed-off-by: Michael Jeronimo <michael.jeronimo@openrobotics.org>