Make the save and record dialogs work the same way
Previously, the code for getting the filename to record a bag, got
a prefix string fom the user and unconditionally appended a suffix
based on the current date/time. In contrast, the code for saving simply
let the user specify the output filename without further modification.

To make these dialogs more uniform in their behavior, this change
prepopulates both dialogs with a filename based on the current date/time.
The user is then free to take the name as-is, or modify it. The result
is then used as the filename for the bag file, resulting in consistent
behavior for Save (export whole bag or region) and Record.

Signed-off-by: Michael Jeronimo <michael.jeronimo@openrobotics.org>