Clean up widgets in plot_view layout correctly
Iterating over the widgets in the layout in forward order when
attempting to delete all of the widgets in a layout resulted in
some widgets not getting deleted properly. Instead, they should
be interated over in reverse order so that the indices for each
widget do not change and all are deleted.

Fixes #5

Signed-off-by: Michael Jeronimo <michael.jeronimo@openrobotics.org>