Refactor the Rosbag2 class
In preparation to switch to rclpy's SequentialWriter, clean up the
Rosbag2 class.

* Merge the rosbag-related utility functions from bag_helper
* Add pydoc comments
* Clean up ament_flake8 issues
* Simplify the SQL query code
* Clean up the interface and remove any redundant functionality

The next step is to use the SequentialWriter, but in order to do
that, we'll need that class to support the kinds of queries we're
doing in this module:

* getting an entry for particular timestamp
* get the entry after a specific timestamp
* get the entries in a range (starting and ending timestamps)

One possibility for this is to extend rosbag2's StorageFilter class to
allow for this kind of filtering.

Signed-off-by: Michael Jeronimo <michael.jeronimo@openrobotics.org>