Revert "Add a dependency on pytest to rqt_bag and rqt_bag_plugins. (#148)"
This reverts commit f0a47cff483b2d776b672c7b5ebb5a1f2c67bde9.

The main reason to do this is that on Windows Debug, we don't currently
have a Debug wheel for PyQt.  Thus these tests always fail on that
platform.  While the "correct" solution is to build a Windows Debug
PyQt wheel, that is a decidedly difficult undertaking.  Since we
don't have tests to speak of in these packages, avoid the problem
for now by just going back to what we were doing before, i.e.
using unittest for the tests.